extract `EditCommentBottomStubPanel` and `EditCommentDefaults`
